// src/app/models/relative.model.ts
export interface Relative {
    relativeId?: string;
    patientId: string;
    relativeName: string;
    relativeFirstLastName: string;
    relativeSecondLastName: string;
    relativeDateOfBirth: Date;
    relativeOccupation: string;
    civilStatus: string;
    relationship: string;
    observations: string;
}